LayoutTests/http/tests/storageAccess/request-storage-access-same-origin-sandboxed-iframe-expected.txt

CONSOLE MESSAGE: line 64: PASS document.hasStorageAccess was granted.
Tests that same-origin iframe storage access is granted if the iframe is sandboxed and has the allow token.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
